Modified tests based on updated classifier architecture
