+++ tests/test_dataset.py
@@ -1,10 +1,7 @@
-import pytest
+import torch
+import networkx as nx
 
-from grace.base import GraphAttrs, Annotation
 from grace.io.image_dataset import ImageGraphDataset
-from grace.models.datasets import dataset_from_graph
-
-from _utils import random_image_and_graph
 
 
 def test_image_graph_dataset(mrc_image_and_annotations_dir):
@@ -16,63 +13,22 @@ def test_image_graph_dataset(mrc_image_and_annotations_dir):
         mrc_image_and_annotations_dir,
         image_filetype="mrc",
     )
-
-    # # all currently fail
-    # image, graph = dataset[0]
-
-    # assert isinstance(image, torch.Tensor)
-    # assert isinstance(graph, nx.Graph)
-
-    # assert image.shape == (1, 128, 128)
-    # assert graph.number_of_nodes() == 4
-
     assert len(dataset) == num_images
 
+    # unwrap sample image & target
+    image, target = dataset[0]
+    graph = target["graph"]
+    metadata = target["metadata"]
+    filename = metadata["image_filename"]
 
-def test_dataset_ignores_subgraph_if_all_edges_unknown(default_rng):
-    _, graph = random_image_and_graph(default_rng)
+    assert isinstance(image, torch.Tensor)
+    assert image.shape == (128, 128)  # 2D image
 
-    edge_update = [
-        (src, dst, {GraphAttrs.EDGE_GROUND_TRUTH: Annotation.UNKNOWN})
-        for src, dst in graph.edges
-    ]
-    graph.update(edges=edge_update)
-    # this action is not currently required since edges are by default UNKNOWN;
-    # however it enables testing of this condition should the default label be changed
+    assert isinstance(metadata, dict)
+    assert isinstance(filename, str)
 
-    assert dataset_from_graph(graph, mode="sub") == []
-
-
-@pytest.mark.parametrize("num_unknown", [7, 17])
-def test_dataset_ignores_subgraph_if_central_node_unknown(
-    num_unknown, default_rng
-):
-    num_nodes_total = 20
-    _, graph = random_image_and_graph(
-        default_rng,
-        num_nodes=num_nodes_total,
-    )
-
-    edge_update = [
-        (src, dst, {GraphAttrs.EDGE_GROUND_TRUTH: Annotation.TRUE_POSITIVE})
-        for src, dst in graph.edges
-    ]
-    graph.update(edges=edge_update)
-
-    node_update = [
-        (node, {GraphAttrs.NODE_GROUND_TRUTH: Annotation.UNKNOWN})
-        for node in list(graph.nodes)[:num_unknown]
-    ]
-    node_update += [
-        (node, {GraphAttrs.NODE_GROUND_TRUTH: Annotation.TRUE_POSITIVE})
-        for node in list(graph.nodes)[num_unknown:]
-    ]
-    graph.update(nodes=node_update)
-
-    assert (
-        len(dataset_from_graph(graph, mode="sub"))
-        == num_nodes_total - num_unknown
-    )
+    assert isinstance(graph, nx.Graph)
+    assert graph.number_of_nodes() == 4
 
 
 def test_dataset_only_takes_common_filenames(tmp_path):
