+++ tests/test_extractor.py
@@ -1,94 +1,11 @@
-from grace.models.classifier import GCNModel
-from grace.models.feature_extractor import FeatureExtractor
-
 import math
 import torch
 import pytest
 
-import networkx as nx
-
-from _utils import random_image_and_graph
-
 from grace.base import GraphAttrs, Annotation
-from grace.models.datasets import dataset_from_graph
+from grace.models.feature_extractor import FeatureExtractor
 
-
-@pytest.mark.parametrize("input_channels", [1, 2])
-@pytest.mark.parametrize("hidden_graph_channels", [[16, 4], [32, 8]])
-@pytest.mark.parametrize("hidden_dense_channels", [[16, 4], [32, 8]])
-@pytest.mark.parametrize("node_output_classes", [2, 4])
-@pytest.mark.parametrize("edge_output_classes", [2, 4])
-class TestGCN:
-    @pytest.fixture
-    def gcn(
-        self,
-        input_channels,
-        hidden_graph_channels,
-        hidden_dense_channels,
-        node_output_classes,
-        edge_output_classes,
-    ):
-        return GCNModel(
-            input_channels=input_channels,
-            hidden_graph_channels=hidden_graph_channels,
-            hidden_dense_channels=hidden_dense_channels,
-            node_output_classes=node_output_classes,
-            edge_output_classes=edge_output_classes,
-        )
-
-    def test_model_building(
-        self,
-        input_channels,
-        hidden_graph_channels,
-        hidden_dense_channels,
-        node_output_classes,
-        edge_output_classes,
-        gcn,
-    ):
-        """Test building the model with different dimensions."""
-
-        assert gcn.conv_layer_list[0].in_channels == input_channels
-
-        assert gcn.node_classifier.in_features == hidden_dense_channels[-1]
-        assert gcn.node_classifier.out_features == node_output_classes
-
-        assert gcn.edge_classifier.in_features == hidden_dense_channels[-1] * 2
-        assert gcn.edge_classifier.out_features == edge_output_classes
-
-    @pytest.mark.parametrize("num_nodes", [4, 5])
-    def test_output_sizes(
-        self,
-        input_channels,
-        hidden_graph_channels,
-        hidden_dense_channels,
-        node_output_classes,
-        edge_output_classes,
-        gcn,
-        num_nodes,
-        default_rng,
-    ):
-        _, graph = random_image_and_graph(
-            default_rng, num_nodes=num_nodes, feature_ndim=input_channels
-        )
-        graph.update(
-            edges=[
-                (
-                    src,
-                    dst,
-                    {GraphAttrs.EDGE_GROUND_TRUTH: Annotation.TRUE_POSITIVE},
-                )
-                for src, dst in graph.edges
-            ]
-        )
-        data = dataset_from_graph(graph, mode="sub")[0]
-
-        subgraph = nx.ego_graph(graph, 0)
-        num_nodes = subgraph.number_of_nodes()
-        num_edges = subgraph.number_of_edges()
-        node_x, edge_x = gcn(x=data.x, edge_index=data.edge_index)
-
-        assert node_x.size() == (num_nodes, node_output_classes)
-        assert edge_x.size() == (num_edges, edge_output_classes)
+from conftest import random_image_and_graph
 
 
 @pytest.mark.parametrize(
@@ -123,7 +40,7 @@ class TestFeatureExtractor:
 
         assert torch.equal(vars["image"], image_out)
 
-        for node_id, node_attrs in graph_out.nodes.data():
+        for _, node_attrs in graph_out.nodes.data():
             x, y = node_attrs[GraphAttrs.NODE_X], node_attrs[GraphAttrs.NODE_Y]
             features = node_attrs[GraphAttrs.NODE_FEATURES]
